+++ sdk/securityinsight/arm-securityinsight/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Release History
     
+## 1.0.0-beta.3 (2022-04-12)
+    
+**Features**
+
+  - Added Type Alias EntityProviders
+  - Type Alias EntityAnalytics has a new parameter entityProviders
+  - Added Enum KnownEntityProviders
+
+**Breaking Changes**
+
+  - Type Alias EntityAnalytics no longer has parameter isEnabled
+  - Type Alias WatchlistItem no longer has parameter itemsKeyValue
+  - Type Alias WatchlistItem no longer has parameter entityMapping
+    
+    
 ## 1.0.0-beta.2 (2022-03-17)
     
 **Features**